+++ pycvvdp/lpyr_dec.py
@@ -1,4 +1,5 @@
 # Decimated Laplacian Pyramid
+from functools import cache
 import torch
 import torch.nn.functional as Func
 import numpy as np 
@@ -131,13 +132,13 @@ class lpyr_dec():
         z = torch.zeros( new_shape, dtype=x.dtype, device=x.device)
         odd_no = (exp_size[dim]%2)
         if dim==-2:
-            z[:,:,2:-2:2,:] = x
-            z[:,:,0,:] = x[:,:,0,:]
-            z[:,:,-2+odd_no,:] = x[:,:,-1,:]
+            z[...,2:-2:2,:] = x
+            z[...,0,:] = x[...,0,:]
+            z[...,-2+odd_no,:] = x[...,-1,:]
         elif dim==-1:
-            z[:,:,:,2:-2:2] = x
-            z[:,:,:,0] = x[:,:,:,0]
-            z[:,:,:,-2+odd_no] = x[:,:,:,-1]
+            z[...,:,2:-2:2] = x
+            z[...,:,0] = x[...,:,0]
+            z[...,:,-2+odd_no] = x[...,:,-1]
         else:
             assert False, "Wrong dimension"
 
@@ -169,16 +170,17 @@ class lpyr_dec():
 
             return torch.cat((beg, x, end), axis)
 
+    @cache
     def get_kernels( self, im, kernel_a = 0.4 ):
 
-        ch_dim = len(im.shape)-2
-        if hasattr(self, "K_horiz") and ch_dim==self.K_ch_dim:
-            return self.K_vert, self.K_horiz
+        # ch_dim = len(im.shape)-2
+        # if hasattr(self, "K_horiz") and ch_dim==self.K_ch_dim:
+        #     return self.K_vert, self.K_horiz
 
         K = torch.tensor([0.25 - kernel_a/2.0, 0.25, kernel_a, 0.25, 0.25 - kernel_a/2.0], device=im.device, dtype=im.dtype)
-        self.K_vert = torch.reshape(K, (1,)*ch_dim + (K.shape[0], 1))
-        self.K_horiz = torch.reshape(K, (1,)*ch_dim + (1, K.shape[0]))
-        self.K_ch_dim = ch_dim
+        self.K_vert = torch.reshape(K, (1, 1, K.shape[0], 1))
+        self.K_horiz = torch.reshape(K, (1, 1, 1, K.shape[0]))
+        # self.K_ch_dim = ch_dim
         return self.K_vert, self.K_horiz
         
 
@@ -186,25 +188,26 @@ class lpyr_dec():
 
         K_vert, K_horiz = self.get_kernels( x, kernel_a )
 
-        B, C, H, W = x.shape
-        y_a = Func.conv2d(x.view(-1,1,H,W), K_vert, stride=(2,1), padding=(2,0)).view(B,C,-1,W)
+        H, W = x.shape[-2], x.shape[-1]
+        y_a = Func.conv2d(x.view(-1,1,H,W), K_vert, stride=(2,1), padding=(2,0)).view( x.shape[0:-2] + (-1,W) )
+        # view(B,C,-1,W)
 
         # Symmetric padding 
-        y_a[:,:,0,:] += x[:,:,0,:]*K_vert[0,0,1,0] + x[:,:,1,:]*K_vert[0,0,0,0]
+        y_a[...,0,:] += x[...,0,:]*K_vert[...,1,0] + x[...,1,:]*K_vert[...,0,0]
         if (x.shape[-2] % 2)==1: # odd number of rows
-            y_a[:,:,-1,:] += x[:,:,-1,:]*K_vert[0,0,3,0] + x[:,:,-2,:]*K_vert[0,0,4,0]
+            y_a[...,-1,:] += x[...,-1,:]*K_vert[...,3,0] + x[...,-2,:]*K_vert[...,4,0]
         else: # even number of rows
-            y_a[:,:,-1,:] += x[:,:,-1,:]*K_vert[0,0,4,0]
+            y_a[...,-1,:] += x[...,-1,:]*K_vert[...,4,0]
 
         H = y_a.shape[-2]
-        y = Func.conv2d(y_a.view(-1,1,H,W), K_horiz, stride=(1,2), padding=(0,2)).view(B,C,H,-1)
+        y = Func.conv2d(y_a.view(-1,1,H,W), K_horiz, stride=(1,2), padding=(0,2)).view( x.shape[0:-2] + (H,-1) )
 
         # Symmetric padding 
-        y[:,:,:,0] += y_a[:,:,:,0]*K_horiz[0,0,0,1] + y_a[:,:,:,1]*K_horiz[0,0,0,0]
+        y[...,:,0] += y_a[...,:,0]*K_horiz[...,0,1] + y_a[...,:,1]*K_horiz[...,0,0]
         if (x.shape[-2] % 2)==1: # odd number of columns
-            y[:,:,:,-1] += y_a[:,:,:,-1]*K_horiz[0,0,0,3] + y_a[:,:,:,-2]*K_horiz[0,0,0,4]
+            y[...,:,-1] += y_a[...,:,-1]*K_horiz[...,0,3] + y_a[...,:,-2]*K_horiz[...,0,4]
         else: # even number of columns
-            y[:,:,:,-1] += y_a[:,:,:,-1]*K_horiz[0,0,0,4] 
+            y[...,:,-1] += y_a[...,:,-1]*K_horiz[...,0,4] 
 
         return y
 
@@ -226,13 +229,13 @@ class lpyr_dec():
 
         y_a = self.interleave_zeros_and_pad(x, dim=-2, exp_size=sz)
 
-        B, C, H, W = y_a.shape
-        y_a = Func.conv2d(y_a.view(-1,1,H,W), K_vert*2).view(B,C,-1,W)
+        H, W = y_a.shape[-2], y_a.shape[-1]
+        y_a = Func.conv2d(y_a.view(-1,1,H,W), K_vert*2).view( x.shape[0:-2] + (-1,W) )
 
         y   = self.interleave_zeros_and_pad(y_a, dim=-1, exp_size=sz)
-        B, C, H, W = y.shape
+        H, W = y.shape[-2], y.shape[-1]
 
-        y   = Func.conv2d(y.view(-1,1,H,W), K_horiz*2).view(B,C,H,-1)
+        y   = Func.conv2d(y.view(-1,1,H,W), K_horiz*2).view( x.shape[0:-2] + (H,-1) )
 
         return y
 
@@ -378,11 +381,8 @@ class weber_contrast_pyr(lpyr_dec):
                 if self.contrast.endswith('ref'):
                     L_bkg = torch.clamp(gpyr[i][...,1:2,:,:,:], min=0.01)
                 else:
-                    L_bkg = torch.clamp(gpyr[i][...,0:2,:,:,:], min=0.01)
                     # The sustained channels use the mean over the image as the background. Otherwise, they would be divided by itself and the contrast would be 1.
-                    L_bkg_mean = torch.mean(L_bkg, dim=[-1, -2], keepdim=True)
-                    L_bkg = L_bkg.repeat([int(image.shape[-4]/2), 1, 1, 1])
-                    L_bkg[0:2,:,:,:] = L_bkg_mean
+                    L_bkg = torch.mean(torch.clamp(gpyr[i][...,0:2,:,:,:], min=0.01), dim=[-1, -2], keepdim=True)
             else:
                 glayer_ex = self.gausspyr_expand(gpyr[i+1], [gpyr[i].shape[-2], gpyr[i].shape[-1]], kernel_a)
                 layer = gpyr[i] - glayer_ex 
@@ -400,8 +400,8 @@ class weber_contrast_pyr(lpyr_dec):
 
             if L_bkg.shape[-4]==2: # If L_bkg NOT identical for the test and reference images
                 contrast = torch.empty_like(layer)
-                contrast[...,0::2,:,:,:] = torch.clamp(torch.div(layer[...,0::2,:,:,:], L_bkg[...,0,:,:,:]), max=1000.0)    
-                contrast[...,1::2,:,:,:] = torch.clamp(torch.div(layer[...,1::2,:,:,:], L_bkg[...,1,:,:,:]), max=1000.0)    
+                contrast[...,0::2,:,:,:] = torch.clamp(torch.div(layer[...,0::2,:,:,:], L_bkg[...,0:1,:,:,:]), max=1000.0)    
+                contrast[...,1::2,:,:,:] = torch.clamp(torch.div(layer[...,1::2,:,:,:], L_bkg[...,1:2,:,:,:]), max=1000.0)    
             else:
                 contrast = torch.clamp(torch.div(layer, L_bkg), max=1000.0)